+++ Template/Trigger/AllLinksClickTrigger.web.js
@@ -9,7 +9,11 @@
                     return nodeName === 'A' || nodeName === 'AREA';
                 }
 
-                TagManager.dom.addEventListener(parameters.document.body, 'click', function (event) {
+                TagManager.dom.onClick(function (event, clickButton) {
+                    clickCallback(event, triggerEvent, clickButton);
+                });
+
+                function clickCallback(event, triggerEvent, clickButton) {
                     if (!event.target) {
                         return;
                     }
@@ -30,10 +34,11 @@
                             'mtm.clickElementClasses': TagManager.dom.getElementClassNames(target),
                             'mtm.clickText': TagManager.dom.getElementText(target),
                             'mtm.clickNodeName': nodeName,
-                            'mtm.clickElementUrl': TagManager.dom.getElementAttribute(target, 'href')
+                            'mtm.clickElementUrl': TagManager.dom.getElementAttribute(target, 'href'),
+                            'mtm.clickButton': clickButton
                         });
                     }
-                }, true);
+                }
 
             });
         };