+++ javascripts/tagmanager.js
@@ -930,6 +930,20 @@
                             callback();
                         }
                     });
+                },
+                onClick: function (callback, element) {
+                    if (typeof element === 'undefined') {
+                        element = documentAlias.body;
+                    }
+                    TagManager.dom.addEventListener(element, 'click', function (event) {
+                        callback(event, 'left')
+                    }, true)
+                    TagManager.dom.addEventListener(element, 'auxclick', function (event) {
+                        callback(event, 'middle')
+                    }, true)
+                    TagManager.dom.addEventListener(element, 'contextmenu', function (event) {
+                        callback(event, 'right')
+                    }, true)
                 }
             };