+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 0.3.6
 
 * Added support to debug a site without appending the debug parameters in the URL
+* Added IntersectionObserver to determine Element Visibility 
+* Added option to observer DOM changes via MutationObserver
 
 0.2.6